Test FTDI read timeout to be 16 ms
On Linux, there is a strange bug that sets the timeout to zero.
However, after sleep, with the FTDI device still plugged in *and
used at least once*, the timeout kicks in and the effective
timeout will be the 16 ms promised by a timeout inside the FTDI
chip.

Test that the timeout is 16 ms, not less.  We expect to not see
data, so the PMT voltages must be zero!