Started work on new HiSPARCIII class
The old Hardware class is a bit of a mess, with self.master, but one
config, potentially multiple buffers, etc. I want two instances of the
new class, one for master, one for slave. They are the ones that
communicate with the device. Above that will be a class which combines
the separate messages. Using that class, you basically don’t have to
worry about master or slave.