Better test for destructor, and fix
We’re now testing the destructor is actually called when the object is
deleted. We needed to keep a weak reference in the config instance, to
make sure __del__ is actually called. Thanks to Eli Bendersky
(http://eli.thegreenplace.net/2009/06/12/safely-using-destructors-in-pyt
hon/)

But: the classes are probably not decoupled enough. I could use the
help of a software architect right now. What is the correct pattern?
Maybe I should have used a more monolithic approach: keep the config in
the HiSPARCIII class itself. However, the HiSPARCIII class strictly
does not need any configuration. You can configure the device by
sending messages. So to place the config in its own class is a good
idea. The problem is how to make sure the config can send the messages.