Turned test_hv script into diagnostic mode
Events are not analyzed, traces are not unpacked, all to keep things fast,
fast, fast!  For each read, a character is written to stdout.  A '.' for
no data, a 'H' for a HiSPARC event and a 'S' for a one-second message.  In
case of a read which takes long, a '!' is written.  After 60 seconds
without data, the script raises a RuntimeError.

This script has been running for about 48 hours without problems.  What is
it that makes the hardware hang sometimes?  It really might be slow reads,
as the script is now much faster and runs without problems.