Initial work on Stew class
Stew event and one-second messages together until the events are cooked.
Be sure to stir regularly and serve the evens when ready.

The thing is, the trigger time as reported by the GPS is not yet corrected
for quantization and synchronization errors, as well as scaled with the
total number of clock ticks in the full second during which the event took
place.  We need a total of three one-second messages to get all the
necessary data.  An event is not ready until up to three seconds after the
event took place.