Dropped lazy evaluation of event attributes
Alas! While it's a nice idea in theory, in practice all attributes were
computed twice: once for each datastore.  The lazy attributes were not
computed before the event was stored in the key-value store.  So when a
datastore needs the value of an attribute for storage, it is computed.
Once, for each datastore.  Sigh.

We could implement some sort of pre-queue.  We store unevaluated events in
the key-value store, where they are picked up by a pool of workers.  Then
all attributes are computed and finally the evaluated event is put in the
queue for storage.  The pool of workers could be implemented using
multiprocessing, resulting in the basic analysis of events using many
cores.  Not on the Pi, of course.