Added acceptance test for ConfigEvents
Proving simply encapsulating the different configs does not work (due
to a weakref to the device which cannot be pickled). This is for the
better: I should not be passing objects around with specific (and
different!) attribute sets which the datastore must unravel. No, the
datastore should receive an event, with attributes matching what will
actually be transmitted and stored.