Track down one second timing offset
WE FOUND IT!

There is a delay between the GPS PPS and the GPS Primary Timing Packet
of up to 20 ms, according to the Trimble docs. This means that the
HiSPARC module can't be sure of the exact GPS time unless it keeps
track of time itself. In that case, it would need to increment the
clock with one second whenever it receives the PPS, but it should tick
over seconds, minutes, hours, days, etc. when needed. It doesn't do
that. When it receives a PrimaryTimingPacket, the HiSPARC hardware sets
it apart and does *not* update the clock. Only when it receives the
next PPS, it shifts in the now *old* PrimaryTimingPacket, sends it to
the software and uses that clock time to use for event messages. That
means that the HiSPARC clock is *behind the GPS clock by exactly one
second!* The hardware relies on the software to rectify that.