Find operating voltage of muonlab
At low voltages, we see no particles. Around the operating voltage, I
expect a plateau when we detect all particles. At high voltages, there
is a high dark current. Limiting factor of this measurement is that we
don’t detect single particles, but only decays of a muon (two
particles).