Increase startup time for PySPARC
This timeout is used to determine of PySPARC exited too quickly. This
happens when the hardware is not present, for example. Setuptools now
uses a lot of time to start up, resulting in pysparc2 keeping
restarting pysparc because it takes longer than 10 sec to start and
figure out there is no hardware present.